Move diff-viewing functionality to own module
